% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_data_concurrent}
\alias{aggregate_data_concurrent}
\title{Aggregate data with concurrent comparators}
\usage{
aggregate_data_concurrent(
  person_data,
  vaccination_data,
  outcome_data,
  period_start,
  period_end,
  options,
  comparator = "vaccinated",
  allow_side_effects = FALSE,
  analysis_types = NULL,
  return_preaggregated = FALSE
)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{vaccination_data}{Vaccination data input data.table. Set to NULL to perform historical data aggregation}

\item{outcome_data}{AESI outcome data input data.table}

\item{period_start}{Start date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{period_end}{End date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}

\item{comparator}{Specify whether a 'vaccinated' or 'unvaccinated' comparator group is used}

\item{allow_side_effects}{Whether or not changes made to the original input data should be allowed to persist after the function has completed.}

\item{analysis_types}{Character vector defining which of the primary and sub-analyses are included in this study (from options file)}

\item{return_preaggregated}{Due to the calculation of ratios and assignment of controls, two fully aggregated tables can not be combined under this study design.
When TRUE, the function will not combine risk/control groups and calculate the ratio, but instead return these objects separately in a list. This allows the majority
of the aggregation to be performed in stages, before combining and calculating the ratio at the end (using \code{concurrent_final_aggregation()} function)}
}
\value{
If \code{return_preaggregated = FALSE}, returns a data.table object matching the 'Concurrent - Vac' or 'Concurrent - Unvac' templates.
If \code{return_preaggregated = TRUE}, returns a list of four data.table objects corresponding to the control/risk counts and control/risk groups, which are required to produce the final table.
}
\description{
Aggregate risk/control AESI counts and ratio according to concurrent study designs, within a given period of time.
}
