% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_data_descriptive_outcomes}
\alias{aggregate_data_descriptive_outcomes}
\title{Aggregate AESI outcomes}
\usage{
aggregate_data_descriptive_outcomes(
  person_data,
  outcome_data,
  period_start,
  period_end,
  options,
  allow_side_effects = FALSE
)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{outcome_data}{AESI outcome data input data.table}

\item{period_start}{Start date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{period_end}{End date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{options}{Options object, already converted from txt file.}

\item{allow_side_effects}{Whether or not changes made to the original input data should be allowed to persist after the function has completed.}
}
\value{
A data.table object matching the 'Descriptive - AESI' template, describing the AESI outcome counts for this period.
}
\description{
Aggregate number of AESI outcomes in a given period of time, regardless of vaccination status.
}
