% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_data_descriptive_vaccinations}
\alias{aggregate_data_descriptive_vaccinations}
\title{Aggregate vaccinations}
\usage{
aggregate_data_descriptive_vaccinations(
  person_data,
  vaccination_data,
  period_start,
  period_end,
  options,
  allow_side_effects = FALSE,
  analysis_types = NULL
)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{vaccination_data}{Vaccination data input data.table. Set to NULL to perform historical data aggregation}

\item{period_start}{Start date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{period_end}{End date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}

\item{allow_side_effects}{Whether or not changes made to the original input data should be allowed to persist after the function has completed.}

\item{analysis_types}{Character vector defining which of the primary and sub-analyses are included in this study (from options file)}
}
\value{
A data.table object matching the 'Descriptive - Vaccinations' template, describing the vaccination counts for this period.
}
\description{
Aggregate number of vaccinations in a given period of time
}
