% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_data_self}
\alias{aggregate_data_self}
\title{Aggregate data with self controls}
\usage{
aggregate_data_self(
  person_data,
  vaccination_data,
  outcome_data,
  period_start,
  period_end,
  options,
  comparator = "post",
  allow_side_effects = FALSE,
  analysis_types = NULL
)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{vaccination_data}{Vaccination data input data.table. Set to NULL to perform historical data aggregation}

\item{outcome_data}{AESI outcome data input data.table}

\item{period_start}{Start date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{period_end}{End date of period, as a date object. Period lengths are defined in the options file, with each cycle possibly containing multiple periods. Only a single period is aggregated at a time.}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}

\item{comparator}{Specify whether a 'post' or 'pre' vaccination control window is used.}

\item{allow_side_effects}{Whether or not changes made to the original input data should be allowed to persist after the function has completed.}

\item{analysis_types}{Character vector defining which of the primary and sub-analyses are included in this study (from options file)}
}
\value{
A data.table object matching the 'SCRI - post' or 'SCRI - pre' templates.
}
\description{
Aggregate risk/control AESI counts and window ratios according to self-controlled study designs, within a given period of time.
}
