% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_clean_window}
\alias{apply_clean_window}
\title{Apply clean window rules}
\usage{
apply_clean_window(data, windows)
}
\arguments{
\item{data}{A data.table of AESI outcome data on which to apply the clean window}

\item{windows}{A data.table of AESI outcome codes and their corresponding clean windows (columns named 'AESI' and 'clean_window')}
}
\value{
A data.table of AESI outcome data with non-incident cases removed
}
\description{
Given a clean window, only the first AESI outcome within that window will be kept.
This is to aid identification of incident outcomes.
}
