% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_vaccine_dose}
\alias{calc_vaccine_dose}
\title{Calculate cumulative vaccine dose numbers}
\usage{
calc_vaccine_dose(data, reference)
}
\arguments{
\item{data}{A data.table of vaccination data, containing at least PID and V_DATE columns, and possibly V_SUBTYPE and V_TYPE depending on specified reference}

\item{reference}{Defines how to accumulate vaccine doses: by 'dose' (all doses counted), 'subtype' (accumulates by subtype), or 'type' (accumulates by type).}
}
\value{
A data.table of vaccination data, with additional column 'V_DOSE' representing cumulative vaccine dose for each person
}
\description{
Calculate cumulative vaccine dose numbers
}
