% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date_format}
\alias{check_date_format}
\title{Check format of date}
\usage{
check_date_format(date)
}
\arguments{
\item{date}{Character, number, or date object, representing a date.}
}
\value{
Date object
}
\description{
Checks that the inputted date is of a recognisable format, and if so, converts
it to a date object, otherwise throws an error
}
