% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_options_object}
\alias{check_options_object}
\title{Check options object}
\usage{
check_options_object(options)
}
\arguments{
\item{options}{Options object in list format}
}
\description{
Checks that the options object has the correct format and contains all the required information.
If not, throws an error.
}
