% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_temp_results}
\alias{combine_temp_results}
\title{Combine temporary results}
\usage{
combine_temp_results(
  design_selection,
  results_file_path,
  options,
  historical_aggregation,
  output_format = "parquet"
)
}
\arguments{
\item{design_selection}{Character vector of specific study designs to run}

\item{results_file_path}{File path to results folder (the main folder created for this cycle)}

\item{options}{Options object, already converted from txt file.}

\item{historical_aggregation}{Boolean stating whether or not these results are for historical aggregation, so that it knows what files to expect.}

\item{output_format}{Output file format: \code{"parquet"} or \code{"csv"}.}
}
\description{
Aggregation is not performed all in one go, but in sections, to manage memory use and allow stopping/restarting.
The results are saved as they are completed, in a 'temp' folder. Once all the sections have been aggregated, the
results are combined using this function, saved in the main results folder, and the 'temp' folder and its contents are removed.
}
