% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{concurrent_final_aggregation}
\alias{concurrent_final_aggregation}
\title{Complete aggregation of data with concurrent comparators}
\usage{
concurrent_final_aggregation(
  control_group,
  risk_group,
  control_counts,
  case_counts,
  output,
  analysis_types,
  options
)
}
\arguments{
\item{control_group}{Table describing people in control group on each day, obtained from \code{aggregate_data_concurrent()} function.}

\item{risk_group}{Table describing people in risk group on each day, obtained from \code{aggregate_data_concurrent()} function.}

\item{control_counts}{Table describing AESI counts from control group on each day, obtained from \code{aggregate_data_concurrent()} function.}

\item{case_counts}{Table describing AESI counts from risk group on each day, obtained from \code{aggregate_data_concurrent()} function.}

\item{output}{A data.table containing already aggregated results, to which these results will be added}

\item{analysis_types}{Character vector defining which of the primary and sub-analyses are included in this study (from options file)}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}
}
\value{
A data.table object matching the 'Concurrent - Vac' or 'Concurrent - Unvac' templates.
}
\description{
Helper function to combine risk/control AESI counts and risk/control groups and calculate ratio, for concurrent study designs.
}
