% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_date_to_number}
\alias{convert_date_to_number}
\title{Convert date to number}
\usage{
convert_date_to_number(date)
}
\arguments{
\item{date}{Date object}
}
\value{
8 digit integer representing a date in the format yyyymmdd
}
\description{
Converts a date object into an 8 digit integer representing a date in the format yyyymmdd.
}
