% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_number_to_date}
\alias{convert_number_to_date}
\title{Convert number to date}
\usage{
convert_number_to_date(date_number)
}
\arguments{
\item{date_number}{8 digit integer representing a date in format yyyymmdd}
}
\value{
Date object
}
\description{
Converts an 8 digit integer 'date' into a date object.
WARNING: does not check that inputted number is in the correct format
}
