% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_and_split_line}
\alias{extract_and_split_line}
\title{Extract data from line of text}
\usage{
extract_and_split_line(title, lines)
}
\arguments{
\item{title}{Name or title of line in text, will search for 'title:' at the start of the line.}

\item{lines}{Character vector containing separate lines of text}
}
\value{
Character vector containing information from line of text
}
\description{
For use in reading data from options file.
Finds line starting with 'title', then extracts everything after the ':' and splits by spaces.
Results are stored in a vector.
}
