% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_first_vac}
\alias{find_first_vac}
\title{Find first vaccination}
\usage{
find_first_vac(data)
}
\arguments{
\item{data}{A data.table of vaccination data, containing at least PID and V_DATE columns}
}
\value{
A data.table of vaccination data, with additional column 'first_vac' representing date of first vaccination for each person
}
\description{
Finds the date of first vaccination for each person, based on the data provided.
}
