% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_person_data}
\alias{generate_person_data}
\title{Generate synthetic person data}
\usage{
generate_person_data(
  pop_size = 1e+06,
  study_start_date = as.Date("2021/01/01"),
  study_end_date = as.Date("2024/01/01"),
  dob_start_date = study_start_date - 100 * 365.25,
  dob_end_date = study_end_date,
  death_age_beta1 = 10,
  death_age_beta2 = 2,
  death_age_max = 100,
  sexes = c("m", "f")
)
}
\arguments{
\item{pop_size}{Number of people in population}

\item{study_start_date}{Date object defining start of the study}

\item{study_end_date}{Date object defining end of the study}

\item{dob_start_date}{Date object defining earliest possible date of birth in cohort (defaults to 100 years before start of study)}

\item{dob_end_date}{Date object defining latest possible date of birth in cohort (defaults to study end date)}

\item{death_age_beta1}{Shape parameter 1 for beta distribution of age at death}

\item{death_age_beta2}{Shape parameter 2 for beta distribution of age at death}

\item{death_age_max}{Maximum age at death}

\item{sexes}{Options for biological sex at birth (drawn uniformly)}
}
\value{
A data.table matching the person data template, with \code{pop_size} rows (BUT date columns are date objects, not in 8-digit number format)
}
\description{
Generates a cohort of size \code{pop_size}, with dates of birth sampled uniform randomly
from between two specified dates (default is between 100 years before the study start and the study end).
Age at death is based on a beta distribution with specified parameters, which defines date of death.
Anyone who dies before the study start is removed, and the whole process is repeated until \code{pop_size} is reached.
}
