% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate synthetic input datasets}
\usage{
generate_synthetic_data(
  pop_size = 1e+06,
  study_start_date = as.Date("2021/01/01"),
  study_end_date = as.Date("2024/01/01"),
  save_data = TRUE,
  save_location = getwd()
)
}
\arguments{
\item{pop_size}{Number of people in population (defines size of person_data)}

\item{study_start_date}{Date object defining start of vaccination campaign and study}

\item{study_end_date}{Date object defining end of study}

\item{save_data}{Boolean whether or not to save the data, as well as return it (default is TRUE, will save)}

\item{save_location}{Location where to save the data, if \code{save_data = TRUE}. Defaults to current working directory}
}
\value{
Named list containing 3 datasets: "person_data", "vaccination_data", and "outcome_data", matching their respective templates.
}
\description{
This function is a simplified wrapper around three more complex functions which generate synthetic
person, vaccination, and AESI outcome data (see \code{generate_person_data()}, \code{generate_vaccination_data()},
and \code{generate_outcome_data()} respectively). Importantly, it also converts the date columns of all three datasets
from Date objects to the 8-digit number format required.

For greater control over the datasets produced, use the 3 specific functions listed above.
}
\examples{
\dontrun{
# Generate synthetic data for 1000 people (without saving to disk)
data <- generate_synthetic_data(
  pop_size = 1000,
  study_start_date = as.Date("2021-01-01"),
  study_end_date = as.Date("2023-01-01"),
  save_data = FALSE
)

head(data$person_data)
head(data$vaccination_data)
head(data$outcome_data)
}

}
