% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_vaccination_data}
\alias{generate_vaccination_data}
\title{Generate synthetic vaccination data}
\usage{
generate_vaccination_data(
  person_data,
  campaign_start_date = as.Date("2021/01/01"),
  campaign_end_date = as.Date("2024/01/01"),
  vaccine_codes = c("ABC", "DEF", "GHI"),
  mean_doses = 3
)
}
\arguments{
\item{person_data}{Person data input data.table (with date columns as Date objects, not 8-digit number format)}

\item{campaign_start_date}{Date object defining start of vaccination campaign (generally same as study start date)}

\item{campaign_end_date}{Date object defining end of vaccination campaign}

\item{vaccine_codes}{Character vector of options for vaccine brand codes}

\item{mean_doses}{Mean number of doses per person during the campaign, as parameter for Poisson distribution}
}
\value{
A data.table matching the vaccination data template (BUT date columns are date objects, not in 8-digit number format)
}
\description{
Generates vaccination data for those people contained in \code{person_data}. The number of
vaccinations each person receives is drawn from a poisson distribution with specified mean
(can be 0). After the number of vaccinations is drawn, vaccination dates and brands are
uniform randomly sampled.
}
