% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_descriptives}
\alias{get_descriptives}
\title{Calculate descriptive statistics}
\usage{
get_descriptives(
  person_data,
  vaccination_data,
  outcome_data,
  age_reference_date
)
}
\arguments{
\item{person_data}{Person data input data.table. If NULL, descriptives for this dataset will be skipped.}

\item{vaccination_data}{Vaccination data input data.table. If NULL, descriptives for this dataset will be skipped.}

\item{outcome_data}{AESI outcome data input data.table. If NULL, descriptives for this dataset will be skipped.}

\item{age_reference_date}{Reference date to calculate age from (generally cycle start)}
}
\value{
Character vector of descriptive statistics
}
\description{
Calculate descriptive statistics
}
