% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_combine_summarise}
\alias{load_combine_summarise}
\title{Load, combine, and summarise temp data}
\usage{
load_combine_summarise(
  file_name_pattern,
  file_path,
  summarise_value,
  summarise_columns
)
}
\arguments{
\item{file_name_pattern}{Pattern to look for in file names. Files with matching pattern will be loaded and combined}

\item{file_path}{Path to where to look for files}

\item{summarise_value}{Name(s) of summary column(s), e.g. COUNT}

\item{summarise_columns}{Names of columns to summarise by}
}
\value{
Combined data table
}
\description{
A helper function to load, combine, and summarise temp data, called from \code{combine_temp_results} function.
Incorporates a catch so that empty data does not cause an error.
}
