% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_subtypes_to_type}
\alias{map_subtypes_to_type}
\title{Map vaccine subtypes to types}
\usage{
map_subtypes_to_type(data, vaccine_mapping)
}
\arguments{
\item{data}{A data.table containing at least a column called V_SUBTYPE}

\item{vaccine_mapping}{A data.table containing columns called V_SUBTYPE and V_TYPE}
}
\value{
A data.table identical to data input, but with an extra column called V_TYPE
}
\description{
Map vaccine subtypes to types
}
