% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidcycler-package.R
\docType{package}
\name{rapidcycler-package}
\alias{rapidcycler}
\alias{rapidcycler-package}
\title{rapidcycler: Rapid Cycle Analysis (RCA) for a Global Network}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of methodologies and functions that are used to perform Rapid Cycle Analysis (RCA), developed for the Global Vaccine Data Network (GVDN). Functionality covers the whole process, from data aggregation and validation, to analysis and result visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://global-vaccine-data-network.github.io/rapidcycler/}
}

}
\author{
\strong{Maintainer}: Benjamin Atkins \email{ben.atkins@mcri.edu.au} (\href{https://orcid.org/0000-0002-3746-8374}{ORCID})

}
\keyword{internal}
