% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_options_file}
\alias{read_options_file}
\title{Read and convert options file}
\usage{
read_options_file(options_file_location)
}
\arguments{
\item{options_file_location}{Location where options.txt file is stored}
}
\value{
Options object in list format
}
\description{
Reads options in from txt file and converts to a list, for use in aggregation functions.
}
