% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_time_periods}
\alias{split_time_periods}
\title{Split time periods}
\usage{
split_time_periods(start, end, period_length, align_periods = FALSE)
}
\arguments{
\item{start}{Start of time interval to split, as a date object}

\item{end}{End of time interval to split, as a date object}

\item{period_length}{Character representing desired length of sub-periods, e.g. 'month', 'day'}

\item{align_periods}{Boolean whether or not to adjust time periods so that they will align,
even if start dates differ. Currently only set up to handle monthly periods, but may be generalised in the future.}
}
\value{
A data.table containing start and end dates of sub-periods
}
\description{
Given a period of time, it will split that interval into sub-periods based on the given length.
}
