% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppress_low_counts}
\alias{suppress_low_counts}
\title{Suppress low counts}
\usage{
suppress_low_counts(data, min_count)
}
\arguments{
\item{data}{A data.table containing data to suppress counts in (will only suppress columns called "COUNT", "CASE_COUNT", "CONTROL_COUNT", or "PERSON_DAYS")}

\item{min_count}{Minimum count allowed - counts less than this will be suppressed, except 0 counts}
}
\value{
A data.table containing data with low counts suppressed
}
\description{
Suppress low counts
}
