% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_column_names}
\alias{validate_column_names}
\title{Validate names of input data columns}
\usage{
validate_column_names(expected, actual, dataset)
}
\arguments{
\item{expected}{Character vector of expected column names}

\item{actual}{Character vector of actual column names}

\item{dataset}{Name of dataset (used in outputted message only)}
}
\value{
A string describing the issue found, or NULL if no issues is found.
}
\description{
Validate names of input data columns
}
