% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_column_types}
\alias{validate_column_types}
\title{Validate names of input data columns}
\usage{
validate_column_types(data, expected, data_name)
}
\arguments{
\item{data}{The dataset being checked}

\item{expected}{Named character vector of expected column types, in the form "COLUMN_NAME" = "COLUMN_TYPE"}

\item{data_name}{Name of dataset (used in outputted message only)}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate names of input data columns
}
