% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_outcome_data}
\alias{validate_outcome_data}
\title{Validate AESI outcome data}
\usage{
validate_outcome_data(outcome_data, person_data, options, split_size)
}
\arguments{
\item{outcome_data}{AESI outcome data input data.table}

\item{person_data}{Person data input data.table}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}

\item{split_size}{Input data will be automatically split into pieces of this size (default 500,000). If memory is an issue, lower the split_size.}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate AESI outcome data
}
