% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_outcome_dates}
\alias{validate_outcome_dates}
\title{Validate outcome dates}
\usage{
validate_outcome_dates(person_data, outcome_data, split_size)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{outcome_data}{AESI outcome data input data.table}

\item{split_size}{Input data will be automatically split into pieces of this size (default 500,000). If memory is an issue, lower the split_size.}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate outcome dates
}
