% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_person_data}
\alias{validate_person_data}
\title{Validate person data}
\usage{
validate_person_data(person_data, cycle_start_date, cycle_end_date)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{cycle_start_date}{Start date of cycle, as a date object, an 8 digit number of the form yyyymmdd, or a string of the form yyyy-mm-dd}

\item{cycle_end_date}{End date of cycle, as a date object, an 8 digit number of the form yyyymmdd, or a string of the form yyyy-mm-dd}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate person data
}
