% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_vaccination_data}
\alias{validate_vaccination_data}
\title{Validate vaccination data}
\usage{
validate_vaccination_data(vaccination_data, person_data, options, split_size)
}
\arguments{
\item{vaccination_data}{Vaccination data input data.table. Set to NULL to perform historical data aggregation}

\item{person_data}{Person data input data.table}

\item{options}{Options object, already converted from txt file. Supply either options_file_location or options, not both}

\item{split_size}{Input data will be automatically split into pieces of this size (default 500,000). If memory is an issue, lower the split_size.}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate vaccination data
}
