% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_vaccination_dates}
\alias{validate_vaccination_dates}
\title{Validate vaccination dates}
\usage{
validate_vaccination_dates(person_data, vaccination_data, split_size)
}
\arguments{
\item{person_data}{Person data input data.table}

\item{vaccination_data}{Vaccination data input data.table. Set to NULL to perform historical data aggregation}

\item{split_size}{Input data will be automatically split into pieces of this size (default 500,000). If memory is an issue, lower the split_size.}
}
\value{
A character vector of issues found. If no issues are found, the vector will be empty.
}
\description{
Validate vaccination dates
}
